# Improve non-kernel scalability of isx-hand-optimized