@ronawho 2018-05-03
@bradcray any objections to this workaround? (note this is isx-hand-optimized, and we'll obviously still want a more principled fix before promoting to the release version)