@bradcray 2018-05-03
OK with me.